Rename DBUnitConfig to BaseDBTest.
BaseDBTest extends BaseTest.
Entity manager runs only once.
