+++ OMS_test/src/test/java/supervisor/EditProductTest.java
@@ -73,7 +73,6 @@ public class EditProductTest extends BaseDBTest {
 
     @After
     public void tearDown() throws Exception {
-        basePage.logout();
         super.tearDown();
     }
 }
